Entities upgrade 0.3.0 (#1337)
* Upgrade to 0.2.0-preview.18
Auto API Upgrade
Time is now on the ComponentSystems
Cleaning up worlds is now done by the package
AssetBundle module is now required

* Change Authority shared component to Authoritative tag component
Add UpdateWorldTimeSystem, required for core
Remove unused WorldsInitializationHelper.cs

* Upgrade to 0.3.0 to resolve burst exception in standalone build

* Changelog

* Update .net to standard 2.0 as required

* Change `Authoritative` to `HasAuthority`

* Generated Code

+++ workers/unity/Packages/io.improbable.gdk.playerlifecycle/Config/PlayerLifecycleConfig.cs
@@ -33,7 +33,7 @@ namespace Improbable.Gdk.PlayerLifecycle
         ///     This is used by the <see cref="SendPlayerHeartbeatRequestSystem"/> on the server-worker to determine
         ///     how often to send a `PlayerHeartbeat` request to client-workers.
         /// </remarks>
-        public static float PlayerHeartbeatIntervalSeconds = 5f;
+        public static double PlayerHeartbeatIntervalSeconds = 5.0;
 
         /// <summary>
         ///     The maximum number of failed heartbeats before a player is disconnected.
